feat: implement authorization

+++ routes/api.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\AuthController;
 use App\Helpers\ApiFormatter;
 use App\Http\Controllers\RoleController;
 use App\Http\Controllers\TraineeController;
+use App\Http\Middleware\RoleMiddleware;
 
 Route::get('/', function () {
     return ApiFormatter::createApi(true, 'This service is working properly', null);
@@ -21,10 +22,10 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/user-by-token', [AuthController::class, 'me']);
 
     Route::get('/users', [UserController::class, 'index']);
-    Route::post('/users', [UserController::class, 'store']);
+    Route::post('/users', [UserController::class, 'store'])->middleware(RoleMiddleware::class . ':admin');
     Route::get('/users/{id}', [UserController::class, 'show']);
     Route::put('/users/{id}', [UserController::class, 'update']);
-    Route::delete('/users/{id}', [UserController::class, 'destroy']);
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware(RoleMiddleware::class . ':admin');
 
     Route::get('/roles', [RoleController::class, 'index']);
 });
